requirements.txt
beautifulsoup4==4.4.1
Pillow==3.0.0
